requirements.txt
pyfemm == 0.1.3